Adjusted `Syntax` type syntax
Tried using `tuple`s for iterables and `list`s for unions.
The problem is that `list`s are not hashable, so it doesn't go well
with generators (which are encoded with `set`s).
The swap makes more sense, since iterables are generally not
hashable, and thus won't be found as the key type for a dict or
something.