Renamed lazy syntax types, and refactored them
The old types are now called `Typed...`. One major issue with them
is that you lose old references (e.g., if a function is to return
a type-checked `list`, then `append`ing to the output does not
append to the original list.
The fixes here do not exactly resolve this, but at least resolves
the issue when the list (for instance) is already `Typed`.