Updated `hash` for $phi$ expressions
`phi` nodes are considered the same (as expressions) if their
arguments coincide; however, `phi` nodes remember their target
as well (to serve as a substitution variable in nested expressions),
but these should hold no bearing on equality of `phi` nodes.
This is true for the `Expr.compare` method, but not for the `hash`
function, which meant that dictionaries could not take advantage
of this.